debug

+++ TelemetryHelper.ps1
@@ -147,29 +147,46 @@ function TrackTrace {
             if ($telemetry.CorrelationId -eq "") {
                 $telemetry.TopId = ""
             }
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "A" }
             $telemetryScope.Properties.Add("Duration", [DateTime]::Now.Subtract($telemetryScope.StartTime).TotalSeconds)
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "B" }
             try {
                 Stop-Transcript | Out-Null
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "C" }
                 $transcript = (@(Get-Content -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)) | select -skip 18 | select -skiplast 4) -join "`n"
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "D" }
                 if ($transcript.Length -gt 30000) {
                     $transcript = "$($transcript.SubString(0,15000))`n`n...`n`n$($transcript.SubString($transcript.Length-15000))"
                 }
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "E $($transcript.Length)" }
                 Remove-Item -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "F" }
             }
             catch {
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "G" }
                 $transcript = ""
             }
 
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "H" }
             $traceTelemetry = $telemetry.Client.GetType().Assembly.CreateInstance('Microsoft.ApplicationInsights.DataContracts.TraceTelemetry')
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "I" }
             $traceTelemetry.Message = "$($telemetryScope.Name)`n$transcript"
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "J" }
             $traceTelemetry.SeverityLevel = $telemetryScope.SeverityLevel
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "K" }
             $telemetryScope.Properties.GetEnumerator() | ForEach-Object { 
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "L $($_.Key) = $($_.Value)" }
                 [void]$traceTelemetry.Properties.TryAdd($_.Key, $_.Value)
             }
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "M" }
             $traceTelemetry.Context.Operation.Name = $telemetryScope.Name
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "N" }
             $traceTelemetry.Context.Operation.Id = $telemetryScope.CorrelationId
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "O" }
             $traceTelemetry.Context.Operation.ParentId = $telemetryScope.ParentId
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "P" }
             $telemetry.Client.TrackTrace($traceTelemetry)
+            if ($telemetry.Debug) { Write-Host -ForegroundColor Yellow "Q" }
             $telemetryScope.Emitted = $true
         }
     }
